refactoring the all json utilities
