+++ autogpt/json_utils/utilities.py
@@ -1,9 +1,30 @@
+"""Utilities for the json_fixes package."""
 import json
+import re
+
 from jsonschema import Draft7Validator
-from autogpt.config import Config
+
 from autogpt.logs import logger
+from autogpt.config import Config
 
 CFG = Config()
+
+def extract_char_position(error_message: str) -> int:
+    """Extract the character position from the JSONDecodeError message.
+
+    Args:
+        error_message (str): The error message from the JSONDecodeError
+          exception.
+
+    Returns:
+        int: The character position.
+    """
+
+    char_pattern = re.compile(r"\(char (\d+)\)")
+    if match := char_pattern.search(error_message):
+        return int(match[1])
+    else:
+        raise ValueError("Character position not found in the error message.")
 
 
 def validate_json(json_object: object, schema_name: object) -> object:
